\name{clear}
\alias{clear}
\title{Deleting Nodes and Relationships}
\description{
Delete all nodes, relationships, constraints, and indexes from the graph database. Requires answering a prompt.
}
\usage{
clear(graph)
}
\arguments{
  \item{graph}{The graph object to be cleared.}
}
\examples{
clear(graph)
}