% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neo4j_http_connection}
\alias{neo4j_http_connection}
\title{Connection to Neo4j}
\usage{
neo4j_http_connection(host, port = 7474)
}
\arguments{
\item{host}{The IP address of the Neo4J instance}

\item{port}{The port to connect to}
}
\description{
Creates a connection object on the host and ports provided
}
\examples{
\dontrun{
   myConn <- neo4j_http_connection(host="localhost");
}
}

