% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neo4j_query_df}
\alias{neo4j_query_df}
\title{Send queries}
\usage{
neo4j_query_df(con, querystring)
}
\arguments{
\item{con}{A neo4j connection object.}

\item{querystring}{The string in double qotes to be passed on to neo4j.}
}
\description{
Send queries to db. Result is a dataframe.
}

