\name{RNeo4j-package}
\alias{RNeo4j}
\title{Interact with a Neo4j database in R.}
\description{
This R package allows you to easily interact with a Neo4j database.
}
\details{
  \tabular{ll}{
  Package: \tab RNeo4j\cr
  Version: \tab 1.0\cr
  Date: \tab 2014-05-02\cr
  }
}
\author{
Nicole White

<nicole@neotechnology.com>
}

