% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNeo4j.R
\name{RNeo4j}
\alias{RNeo4j}
\title{Neo4j Driver for R}
\description{
Neo4j's R driver. RNeo4j allows you to read and write data from / to Neo4j directly from your R environment. For bug reports, please
\href{https://github.com/nicolewhite/RNeo4j/issues}{create a new GitHub issue}.
For support, please \href{http://stackoverflow.com/questions/tagged/r-neo4j}{create a question on StackOverflow}.
}
\author{
Nicole White
}

