% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabel.R
\name{addLabel}
\alias{addLabel}
\title{Node Labels}
\usage{
addLabel(node, ...)
}
\arguments{
\item{node}{A node object.}

\item{...}{A character vector.}
}
\description{
Add a label or multiple labels to an existing node object.
}
\examples{
\dontrun{
graph = startGraph("http://localhost:7474/db/data/")
clear(graph)

alice = createNode(graph, name = "Alice")
bob = createNode(graph, name = "Bob")

addLabel(alice, "Person")
addLabel(bob, c("Person", "Student"))

alice
bob
}

}
\seealso{
\code{\link{getLabel}}, \code{\link{dropLabel}}
}

