% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear the Database}
\usage{
clear(graph, input = TRUE)
}
\arguments{
\item{graph}{A graph object.}

\item{input}{A logical constant. If \code{FALSE}, do not require user input for confirmation.}
}
\description{
Delete all nodes, relationships, constraints, and indexes from Neo4j.
}
\examples{
\dontrun{
graph = startGraph("http://localhost:7474/db/data/")
clear(graph)
}

}

