% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\title{Internal IDs}
\usage{
getID(entity)
}
\arguments{
\item{entity}{A node or relationship object.}
}
\value{
An integer.
}
\description{
Retrieve the internal ID of a node or relationship object.
}
\examples{
\dontrun{
graph = startGraph("http://localhost:7474/db/data/")
clear(graph)

alice = createNode(graph, "Person", name = "Alice")
getID(alice)
}

}

