% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSample.R
\name{importSample}
\alias{importSample}
\title{Import Sample Datasets}
\usage{
importSample(graph, data, input = TRUE)
}
\arguments{
\item{graph}{A graph object.}

\item{data}{A character string. Datasets include "tweets", "dfw", "caltrain", and "movies".}

\item{input}{A logical constant. If \code{TRUE}, require confirmation.}
}
\description{
Populate the graph database with one of the sample datasets supplied with this package.
}
\examples{
\dontrun{
graph = startGraph("http://localhost:7474/db/data/")
importSample(graph, "tweets")
}

}

