\name{RNeo4j-package}
\alias{RNeo4j}
\title{Interact with a Neo4j database in R.}
\description{
An R package that allows you to easily interact with a Neo4j database from your R environment.
}
\details{
  \tabular{ll}{
  Package: \tab RNeo4j\cr
  Version: \tab 1.2.1\cr
  Date: \tab 2014-05-02\cr
  }
}
\author{
Nicole White

<nicole@neotechnology.com>

Please take your questions to StackOverflow and use the r-neo4j tag:

http://stackoverflow.com/questions/tagged/r-neo4j
}

