\name{Rneo4j-package}
\alias{Rneo4j}
\title{Interact with a Neo4j database in R}
\description{
This R package allows you to easily populate and interact with a Neo4j database. It is intended for prototyping smaller datasets. It is also helpful if you are looking to frequently pull data from a Neo4j database into R data frames for analysis or plotting.
}
\details{
  \tabular{ll}{
  Package: \tab Rneo4j\cr
  Version: \tab 1.0\cr
  Date: \tab 2014-05-02\cr
  }
}
\author{
Nicole White

<nicole.m.white@outlook.com>
}

