\name{RNeo4j-package}
\alias{RNeo4j}
\title{Neo4j Driver for R}
\description{
RNeo4j is Neo4j's R driver. It allows you to read and write data from / to Neo4j directly from your R environment.
}
\details{
  \tabular{ll}{
  Package: \tab RNeo4j\cr
  Version: \tab 1.4.2\cr
  Date: \tab 2014-05-02\cr
  }
}
\author{
Nicole White

If you think you've found a bug, please \href{https://github.com/nicolewhite/RNeo4j/issues}{create a new GitHub issue}. If you have a question about how to do something in RNeo4j or need help troubleshooting, please \href{http://stackoverflow.com/questions/tagged/r-neo4j}{post to StackOverflow and use the r-neo4j tag}.

For anything else, you can email me:

<nicole@neotechnology.com>
}

