\name{RNeo4j-package}
\alias{RNeo4j}
\title{Interact with a Neo4j database in R.}
\description{
An R package that allows you to easily interact with a Neo4j database from your R environment.
}
\details{
  \tabular{ll}{
  Package: \tab RNeo4j\cr
  Version: \tab 1.4.1\cr
  Date: \tab 2014-05-02\cr
  }
}
\author{
Nicole White

If you think you've found a bug, please create a new GitHub issue:

https://github.com/nicolewhite/RNeo4j/issues

If you have a question about how to do something in RNeo4j or need help troubleshooting, please post to StackOverflow and use the r-neo4j tag:

http://stackoverflow.com/questions/tagged/r-neo4j

For anything else, you can email me:

<nicole@neotechnology.com>
}

