% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGraph.R
\name{startGraph}
\alias{startGraph}
\title{Connect to the Database}
\usage{
startGraph(url, username = character(), password = character(),
  opts = list())
}
\arguments{
\item{url}{A character string.}

\item{username}{A character string. If authentication is enabled, your username.}

\item{password}{A character string. If authentication is enabled, your password.}

\item{opts}{A named list. Optional HTTP settings.}
}
\value{
A graph object.
}
\description{
Establish a connection to Neo4j.
}
\examples{
\dontrun{
graph = startGraph() # http://localhost:7474/db/data/ by default

graph = startGraph("http://localhost:7474/db/data/", 
                   username = "neo4j",
                   password = "password")

graph = startGraph("http://localhost:7474/db/data/", 
                   opts = list(timeout=3))
}

}

